20190217 Fixes
Fixed issue where memory was not being reported accurately - conky.conf > no_buffers = true. Memory was being reported as being used more due to it including system buffers  that were not actively in use.
Added functionality so that the planets will only turn transparent when they are within +/-15 degrees of the top of the arc, where they might block the text. 
Changed text color of the TOP CPU chart for readability.
Plugged the hole in the center of the "Sun" for readability.
Changed the way the Lua script is referenced in Conky.conf to no longer be a absolute path.